byLocal first draft
By local works, though with major drawbacks.  Also added a "pluginTypes" .h file that all plugins need to be "registered" in, so that the dataStructure knows they exist at compile time for the templated classes that can work with them.  This will all be auto generated at a later point in time.